+++ rust/operator-binary/src/container.rs
@@ -48,6 +48,7 @@ use stackable_operator::{
             CustomContainerLogConfig,
         },
     },
+    role_utils::RoleGroupRef,
     utils::{COMMON_BASH_TRAP_FUNCTIONS, cluster_info::KubernetesClusterInfo},
 };
 use strum::{Display, EnumDiscriminants, IntoStaticStr};
@@ -216,24 +217,25 @@ impl ContainerConfig {
         hdfs: &v1alpha1::HdfsCluster,
         cluster_info: &KubernetesClusterInfo,
         role: &HdfsNodeRole,
-        role_group: &str,
+        rolegroup_ref: &RoleGroupRef<v1alpha1::HdfsCluster>,
         resolved_product_image: &ResolvedProductImage,
         merged_config: &AnyNodeConfig,
         env_overrides: Option<&BTreeMap<String, String>>,
         zk_config_map_name: &str,
-        object_name: &str,
         namenode_podrefs: &[HdfsPodRef],
         labels: &Labels,
     ) -> Result<(), Error> {
         // HDFS main container
         let main_container_config = Self::from(*role);
-        pb.add_volumes(main_container_config.volumes(merged_config, object_name, labels)?)
+        let object_name = rolegroup_ref.object_name();
+
+        pb.add_volumes(main_container_config.volumes(merged_config, &object_name, labels)?)
             .context(AddVolumeSnafu)?;
         pb.add_container(main_container_config.main_container(
             hdfs,
             cluster_info,
             role,
-            role_group,
+            rolegroup_ref,
             resolved_product_image,
             zk_config_map_name,
             env_overrides,
@@ -277,6 +279,8 @@ impl ContainerConfig {
                         )
                         .with_pod_scope()
                         .with_node_scope()
+                        // To scrape metrics behind TLS endpoint (without FQDN)
+                        .with_service_scope(rolegroup_ref.rolegroup_metrics_service_name())
                         .with_format(SecretFormat::TlsPkcs12)
                         .with_tls_pkcs12_password(TLS_STORE_PASSWORD)
                         .with_auto_tls_cert_lifetime(
@@ -319,7 +323,7 @@ impl ContainerConfig {
                 let zkfc_container_config = Self::try_from(NameNodeContainer::Zkfc.to_string())?;
                 pb.add_volumes(zkfc_container_config.volumes(
                     merged_config,
-                    object_name,
+                    &object_name,
                     labels,
                 )?)
                 .context(AddVolumeSnafu)?;
@@ -327,7 +331,7 @@ impl ContainerConfig {
                     hdfs,
                     cluster_info,
                     role,
-                    role_group,
+                    rolegroup_ref,
                     resolved_product_image,
                     zk_config_map_name,
                     env_overrides,
@@ -340,7 +344,7 @@ impl ContainerConfig {
                     Self::try_from(NameNodeContainer::FormatNameNodes.to_string())?;
                 pb.add_volumes(format_namenodes_container_config.volumes(
                     merged_config,
-                    object_name,
+                    &object_name,
                     labels,
                 )?)
                 .context(AddVolumeSnafu)?;
@@ -348,7 +352,7 @@ impl ContainerConfig {
                     hdfs,
                     cluster_info,
                     role,
-                    role_group,
+                    &rolegroup_ref.role_group,
                     resolved_product_image,
                     zk_config_map_name,
                     env_overrides,
@@ -362,7 +366,7 @@ impl ContainerConfig {
                     Self::try_from(NameNodeContainer::FormatZooKeeper.to_string())?;
                 pb.add_volumes(format_zookeeper_container_config.volumes(
                     merged_config,
-                    object_name,
+                    &object_name,
                     labels,
                 )?)
                 .context(AddVolumeSnafu)?;
@@ -370,7 +374,7 @@ impl ContainerConfig {
                     hdfs,
                     cluster_info,
                     role,
-                    role_group,
+                    &rolegroup_ref.role_group,
                     resolved_product_image,
                     zk_config_map_name,
                     env_overrides,
@@ -385,7 +389,7 @@ impl ContainerConfig {
                     Self::try_from(DataNodeContainer::WaitForNameNodes.to_string())?;
                 pb.add_volumes(wait_for_namenodes_container_config.volumes(
                     merged_config,
-                    object_name,
+                    &object_name,
                     labels,
                 )?)
                 .context(AddVolumeSnafu)?;
@@ -393,7 +397,7 @@ impl ContainerConfig {
                     hdfs,
                     cluster_info,
                     role,
-                    role_group,
+                    &rolegroup_ref.role_group,
                     resolved_product_image,
                     zk_config_map_name,
                     env_overrides,
@@ -462,7 +466,7 @@ impl ContainerConfig {
         hdfs: &v1alpha1::HdfsCluster,
         cluster_info: &KubernetesClusterInfo,
         role: &HdfsNodeRole,
-        role_group: &str,
+        rolegroup_ref: &RoleGroupRef<v1alpha1::HdfsCluster>,
         resolved_product_image: &ResolvedProductImage,
         zookeeper_config_map_name: &str,
         env_overrides: Option<&BTreeMap<String, String>>,
@@ -481,7 +485,7 @@ impl ContainerConfig {
             .args(self.args(hdfs, cluster_info, role, merged_config, &[])?)
             .add_env_vars(self.env(
                 hdfs,
-                role_group,
+                &rolegroup_ref.role_group,
                 zookeeper_config_map_name,
                 env_overrides,
                 resources.as_ref(),
@@ -1249,16 +1253,18 @@ wait_for_termination $!
     /// Container ports for the main containers namenode, datanode and journalnode.
     fn container_ports(&self, hdfs: &v1alpha1::HdfsCluster) -> Vec<ContainerPort> {
         match self {
-            ContainerConfig::Hdfs { role, .. } => hdfs
-                .ports(role)
-                .into_iter()
-                .map(|(name, value)| ContainerPort {
-                    name: Some(name),
-                    container_port: i32::from(value),
-                    protocol: Some("TCP".to_string()),
-                    ..ContainerPort::default()
-                })
-                .collect(),
+            ContainerConfig::Hdfs { role, .. } => {
+                // data ports
+                hdfs.hdfs_main_container_ports(role)
+                    .into_iter()
+                    .map(|(name, value)| ContainerPort {
+                        name: Some(name),
+                        container_port: i32::from(value),
+                        protocol: Some("TCP".to_string()),
+                        ..ContainerPort::default()
+                    })
+                    .collect()
+            }
             _ => {
                 vec![]
             }